requirements.txt
akshare==1.18.13
json_repair==0.54.3
loguru==0.7.3
Markdown==3.10
markitdown==0.1.4
openai==2.15.0
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
ta_lib==0.6.4
tavily==1.1.0
